ISPN-5228 Remove redundant code in locking interceptors * GetKeyValueCommand and GetCacheEntryCommand are handled in visitDataReadCommand() * PutKeyValueCommand, RemoveCommand and ReplaceCommand are handled in visitDataWriteCommand()
