+++ core/src/main/java/org/infinispan/interceptors/base/CommandInterceptor.java
@@ -3,7 +3,7 @@ package org.infinispan.interceptors.base;
 import org.infinispan.Cache;
 import org.infinispan.cache.impl.CacheImpl;
 import org.infinispan.commands.AbstractVisitor;
-import org.infinispan.commands.FlagAffectedCommand;
+import org.infinispan.commands.LocalFlagAffectedCommand;
 import org.infinispan.commands.VisitableCommand;
 import org.infinispan.commands.Visitor;
 import org.infinispan.commands.read.GetKeyValueCommand;
@@ -111,7 +111,7 @@ public abstract class CommandInterceptor extends AbstractVisitor {
       return invokeNextInterceptor(ctx, command);
    }
 
-   protected final long getLockAcquisitionTimeout(FlagAffectedCommand command, boolean skipLocking) {
+   protected final long getLockAcquisitionTimeout(LocalFlagAffectedCommand command, boolean skipLocking) {
       if (!skipLocking)
          return command.hasFlag(Flag.ZERO_LOCK_ACQUISITION_TIMEOUT) ?
                0 : cacheConfiguration.locking().lockAcquisitionTimeout();
@@ -119,7 +119,7 @@ public abstract class CommandInterceptor extends AbstractVisitor {
       return -1;
    }
 
-   protected final boolean hasSkipLocking(FlagAffectedCommand command) {
+   protected final boolean hasSkipLocking(LocalFlagAffectedCommand command) {
       return command.hasFlag(Flag.SKIP_LOCKING);
    }
