+++ core/src/main/java/org/infinispan/interceptors/locking/NonTransactionalLockingInterceptor.java
@@ -1,13 +1,10 @@
 package org.infinispan.interceptors.locking;
 
 import org.infinispan.InvalidCacheUsageException;
-import org.infinispan.commands.read.AbstractDataCommand;
-import org.infinispan.commands.read.GetCacheEntryCommand;
-import org.infinispan.commands.read.GetKeyValueCommand;
+import org.infinispan.commands.DataCommand;
 import org.infinispan.commands.write.ClearCommand;
+import org.infinispan.commands.write.DataWriteCommand;
 import org.infinispan.commands.write.PutMapCommand;
-import org.infinispan.commands.write.RemoveCommand;
-import org.infinispan.commands.write.ReplaceCommand;
 import org.infinispan.context.InvocationContext;
 import org.infinispan.util.logging.Log;
 import org.infinispan.util.logging.LogFactory;
@@ -28,21 +25,19 @@ public class NonTransactionalLockingInterceptor extends AbstractLockingIntercept
    }
 
    @Override
-   public final Object visitGetKeyValueCommand(InvocationContext ctx, GetKeyValueCommand command) throws Throwable {
-      return visitDataReadCommand(ctx, command);
-   }
-   @Override
-   public final Object visitGetCacheEntryCommand(InvocationContext ctx, GetCacheEntryCommand command) throws Throwable {
-      return visitDataReadCommand(ctx, command);
-   }
-
-   private Object visitDataReadCommand(InvocationContext ctx, AbstractDataCommand command) throws Throwable {
+   protected final Object visitDataReadCommand(InvocationContext ctx, DataCommand command) throws Throwable {
       assertNonTransactional(ctx);
       try {
          return invokeNextInterceptor(ctx, command);
       } finally {
          lockManager.unlockAll(ctx);//possibly needed because of L1 locks being acquired
       }
+   }
+
+   @Override
+   protected Object visitDataWriteCommand(InvocationContext ctx, DataWriteCommand command) throws Throwable {
+      assertNonTransactional(ctx);
+      return visitNonTxDataWriteCommand(ctx, command);
    }
 
    @Override
@@ -74,37 +69,6 @@ public class NonTransactionalLockingInterceptor extends AbstractLockingIntercept
                   lockKey(ctx, key, lockTimeout, skipLocking);
             }
          }
-         return invokeNextInterceptor(ctx, command);
-      } catch (Throwable te) {
-         throw cleanLocksAndRethrow(ctx, te);
-      }
-      finally {
-         lockManager.unlockAll(ctx);
-      }
-   }
-
-   @Override
-   public Object visitRemoveCommand(InvocationContext ctx, RemoveCommand command) throws Throwable {
-      assertNonTransactional(ctx);
-      try {
-         if (!shouldLock(command.getKey(), command))
-            return invokeNextInterceptor(ctx, command);
-         lockKey(ctx, command);
-         return invokeNextInterceptor(ctx, command);
-      } catch (Throwable te) {
-         throw cleanLocksAndRethrow(ctx, te);
-      } finally {
-         lockManager.unlockAll(ctx);
-      }
-   }
-
-   @Override
-   public Object visitReplaceCommand(InvocationContext ctx, ReplaceCommand command) throws Throwable {
-      assertNonTransactional(ctx);
-      try {
-         if (!shouldLock(command.getKey(), command))
-            return invokeNextInterceptor(ctx, command);
-         lockKey(ctx, command);
          return invokeNextInterceptor(ctx, command);
       } catch (Throwable te) {
          throw cleanLocksAndRethrow(ctx, te);
